refactor: update sessionFlowName import paths and remove unused sessionFlowName constants
